[VSPC-208] addressed review comments
